Fix: Move QrCode
