+++ src/components/hackeps/Profile/profile.js
@@ -19,7 +19,7 @@ import qrIcon from "src/icons/qr.png";
 import Team from "src/components/hackeps/Team/Team";
 import LinkAccounts from "src/components/hackeps/LinkAccounts/LinkAccounts";
 import Join from "src/components/hackeps/Join/Join";
-import QrCode from "src/components/hackeps/Home/QrCode.js";
+import QrCode from "src/components/hackeps/QrCode/QrCode.js";
 import { getHackerGroupById } from "src/services/HackerGroupService";
 import UserNotFound from "./UserNotFound";
 import ProfilePic from "../ProfilePic/ProfilePic";
