+++ src/components/hackeps/QrCode/QrCode.css
